RNG added.
dSFMT: an extremely fast Mersenne Twister from the original authors of
Mersenne Twister.  Implemented in C.  A wrapper to selected functions
is written in C++ to make the functions accessible from Fortran.

In addition there's a helpful interface module in Fortran which makes
accessing a "store" of random numbers very simple.

Note that generating an array of random numbers is much more efficient than
generating them one at a time.