Fix compression of spawned walker list.  Results change.
Previously the compression of the spawned walker list (i.e. annihilation
and compression to the list of unique determinants within the spawned
array) caused the population of the last particle in the spawned list to
be occasionally wrong.

This is a complete rewrite of the algorithm which fixes it and is much
simpler to boot.

I don't know what I was thinking the first time round...