Bug fix to annihilation.
Previously when the spawned list was compressed (such that there's
one entry per determinant with the total population spawned on that
determinant), determinants with a total zero spawned population were
being kept in the spawned list.

This meant that occasionally spawned walkers with a 0 population are
inserted into the main walker list.  This is undesirable.

Results change (but remain within error bars) as walkers with a 0
population consumed random numbers in the death procedure.  Removing
this issue causes simulations to follow different stocastic paths.