Ensure the main walker list doesn't contain any unoccupied determinants.
There were some bug/"features" in the previous annihilation process which allowed
the main walker list to have some unoccupied determinants in it for
a while (they were eventually removed!)

This was because

a) if all the walkers on a determinant died and no walkers were spawned,
the main list was not being compressed.
b) if the last determinant in the spawned list was completely annihilated, then
that was not being removed from the spawned list and so a null
population could be inserted into the main list.

These are both fixed.

Results change (but not within statistical errors) as random numbers are
not being consumed attempting to kill a null population and so the path
through our space changes.