Fix pgen for real space Hubbard model.
The counting of the number of possible excitations was previously incorrect.

A similar fix for the spawning module was introduced in
f03c17cbd3f8a432bca838077d255ce9ecf270f9.

ieor(f, connected_orbs) gives the orbitals occupied in the determinants
not connected to the orbital of interest and vice versa.  We actually
only want the orbitals connected to the orbital of interest and not
occupied in the determinant (i.e. the latter) hence the need for
iand(not(f), connected_orbs).