Subtle fix to create_excited_det_complete.
If a variable of a defined type is set to be intent(out) in a procedure,
then that variable is (re-)initialized when that procedure is called.  A
consequence of this is that any allocatable/pointer attributes are
deallocated/nullified and the procedure is expected to define them.

We actually want create_excited_det_complete to set the attributes to an existing
det_info variable, which has already had its attributes allocated.  Hence cdet_out
must be defined as intent(inout).