Store data rather than the determinant label in det_info type.
Add det_info member which point to relevant 1D subarray in walker_data.
Use this in the importance sampling routines to avoid carrying around
the idet label.

This is tidier and allows the routines to be more flexible.

Coming soon: change interface to the death routines to the
update_proj_energy_* routines to also not use idet.