Rework energy estimator communications.
Certainly one of the most embarrassing parts of the codebase!

At the cost of an (entirely negligible) amount of CPU time, combine the
serial and parallel codes.  As a bonus, this makes adding to the
estimator code much easier in future, as one must only change one
section rather than two!