Correct horrendous bug in MC sampling of the Hilbert space.
The orbitals were not being selected in a uniform fashion.  In
particular, there was a bias in selecting the first two orbitals.  This
created a bias, especially prevalent at lower excitation levels.

This did not come to light until testing with the truncated space---the
full space is usually sufficiently large (and so many more states at the
higher excitation levels) that the biases were not noticeable.

Exact estimates of the hilbert space will change (but the new numbers
are correct at least!).

Ooops!