Another HFS initiator bug fix.
The test of the parent flag when an spawned particle is being created on
a determinant previously only occupied by particles of a different type
was looking at the wrong bit.

HFS with and without the initiator approximation now agree when
initiator_population is set to 0 (a limit which recovers the standard
algorithm).