Update tools/mkconfig.py.
Option to handle the case when f90_module_flag needs to be appended with a space.

Sadly ConfigParser cannot be configured to not strip whitespace and some
Fortran compilers absolutely require a space between the f90_module_flag
and the directory argument whilst other Fortran compilers absolutely
require no space.