Do not normalise the sampling of the reference population.
It seems that, despite 1434bb, we do not need to normalise the reference
population by the number of sampling events.  I don't quite understand
this yet but it's clear that the normalisation introduced in 1434bb
caused the reference population to be too small by exactly the
normalisation factor!  Need to discuss with AJWT.