BUG FIX: freezing core orbitals from a RHF calculation.
Previously we were not taking the various spin permutations correctly
into account when calculating the contribution from the frozen core
orbitals to the core energy.  This was not noticed previously as only
a small number of electrons (ie 2) were frozen or UHF systems where
studied.

WARNING: previously obtained results for affected calculations are
incorrect.  Happily the results will be incorrect by a constant value
(the difference between the wrong and correct values of the core
energy), so any existing results can easily be corrected.