BUG FIX: dumping out restart files in ascii format in parallel.
The psip information on the root processor is overwritten with
information from other processors and so has to be read back in.
This was broken when using ascii format as it tried to parse the restart
file and entered an infinite loop.

Instead, we'll just do what we do when writing restart files in binary:
also write the root information to a scratch file which is then easy to
read back in.

The restart code is horribly fragile (oh, how I regret it!).  The sooner
we move to using HDF5 the better!