Set CC to do Full CC by default.
The truncation level is not set to an initial value normally as it's
only used elsewhere if a truncation level is specified by the user.
However, it is used in selecting the size of a random cluster in the
CCMC code and referencing an uninitialised variable is a little(!) bad.

So, let's do the sane thing and assume the user wants to allow all
possible excitations in CCMC unless they say otherwise.