BUG FIX: normalise the amplitudes correctly.
Previously we were normalising the amplitude by dividing through by N_0^{s-1},
where N_0 is the (integer) population on the reference and s is the cluster
size. This caused integer overflows quite quickly (e.g. if N_0 is ~5000 and
s is 4!).  Promoting N_0 to a real before raising it to the power of s-1 is
much safer...

TODO: we should probably detect for whether a overflow of the real type is
possible...