Zero integral stores before using them in read_in.
This avoids the case where an integral can by non-zero by the (possibly
incomplete) symmetry provided but is in fact zero.  As we don't know the
integral is zero, we might access it and hence read from unitialised memory
(VERY BAD).  Zeroing the arrays is a minor overhead and worth it to make sure
we avoid this problem.