BUG FIX: excitation generation probabilties for no_renorm.
Only Hubbard model in momentum space was affected. See comments in
excit_gen_hub_k for details.  Essentially because we always select an
unoccupied alpha orbital first in this system, there's an asymmetry in
the generation which must be taken into account in the probability of
selecting a given excitation.