Make projected energy update procedures pure (+bonus bug fix).
They now update the D0 population and projected energy denominator sum
via arguments rather than module-level variables, which allows for
OpenMP CCMC code to treat them as reductions rather than requiring
critical blocks.

NOTE: this fixes a bug (uncovered by the purification process) where the
Neel trial function was being initialised to 1 only on the first call
and then subsequently re-used with the value at the end of the previous
call.  This only occurred if the Neel single state was used as the trial
function but importance sampling was not used.  This is the danger with
Fortran's implicit save...