Generic structure for spawning data.
Replace the code for storing and manipulating spawned particles with
a generic structure and routines which operate on that structure.  This
allows us to have multiple, separate object for storing spawned
particles in, e.g. the standard spawning in DMQMC and also accumulation
of a RDM, which uses a different Hilbert space of basis functions.

NOTE: we have disabled reading in restart files for parallel code.  It's
getting increasingly more painful to maintain and needs to be replaced
with a scalable, sane HDF5 solution....

TODO: don't store qmc_data in fciqmc_data but instead have it as a local
instance in the relevant QMC main procedures.

Yay!  More purity!