Move operator overloading of bit_str_gt to bit_utils.
It seems safest if we define the operator interface in the same module
as the procedure is defined.  gfortran (but not ifort) chokes on

    interface operator(...)
        module procedure ...
    end interface

and ifort (but not gfortran) on

    interface operator(...)
        procedure ...
    end interface

when referring to procedures defined in a different module.  The first
version, if given in the same module the procedure is defined in, is
standard Fortran 90 and so supported by all compilers we'll use.