Fix long-standing bug in the initial status of a QMC simulation
There was a (commented!) assumption that the bit string did not need to
be decoded when calculating the projected energy for the initial set of
particles.  This is not true for molecular systems so add in an explicit
decode call.  Not sure how this one avoided notice for so long...