Set include flags for HDF5 for gfortran configs.
These work on the Arch Linux laptop and on the CMTH cluster.  Feel free
to add include paths of your own or your own config files.  Perhaps the
easiest way though is to set HDF5_ROOT to point to the HDF5 installation
(done automatically on the CMTH machines by the module files there, of
course).