Set max cluster size and position of D0 in parallel.
Note that we set the position of D0 in the main list to be -1 if the
processor does not contain the reference, so that we don't need to
change any code in select_cluster to accommodate the absence of D0 in
the cumulative population list.

This comit also contains a small optimisation: use the previous value of
D0_pos to provide a lower or uppper bound to its current position.