Minor CCMC MPI improvement/modification.
Set the number of spawning attempts to be at least the population on the
reference.  This avoids the case where, early in the simulation,
selecting the reference on a processor which does not contain the
reference (and hence has a small population) leads to a large spawning
event (thus coarsening the excip growth/diffusion) as the cluster
amplitude and selection probabilities are not well balanced.