Set default seed to be hash of time and UUID.
Note that for convenience we get the time as a string rather than an
integer array so it can be simply concatenated with the UUID.

Test inputs updated to use the old default (7).  Naturally existing
results will change (stochastically/not within error bars) if input
files are not updated...

This allows one to run a set of calculations where each calculation will
get a different seed automatically.  Much less painful!

Of course, there's still no guarantee that the streams of different
numbers will be independent/uncorrelated but it's the best we can do
without knowing all other seeds used for studying a given system!