write_array wrapper procedures to write arrays to an HDF5 file.
As suggested by Alex Thom, this is to ensure that we call HDF5 with the correct kind/type
parameter for a given array.  This is now guaranteed if using the write_array procedure
(due to type safety).  Previously the programmer had to use write_ptr directly, which
required each call to be given the correct kind/type (ie more error prone).