Deterministically select a reference determinant from a degenerate set.
The automatic selection of a reference determinant (given a symmetry and
spin polarisation) attempts to find the lowest energy determinant.  This
was being done by a strict comparison, which left us vulnerable to
numerical noise amongst members of a degenerate set and hence could give
non-deterministic results.  Using a more relaxed tolerance means that we
now deterministically select the first member of a degenerate set that
we encounter.

Reported by: Fionn Malone.