Bloom threshold for CCMC with MPI parallelisation.
We allow MPI processes in CCMC to spawn from the reference even if the process has no
population on it.  In such cases we should compare to the number of attempts made
before declaring a bloom has occurred.

This removes the spurious 'a bloom has occurred' warnings, even when it results in only
one psip being produced.