Change RNG seeds in CCMC for hybrid parallelisation.
Rather than setting the seeds to be consecutive across threads within the same MPI
process, assign seeds in a round-robin fashion (i.e. process 0 gets the original seed, S,
S+nthreads, S+2nthreads, ...).  This means that the seed given to the first thread within
each MPI process is independent of the number of threads being used.

A minor change but it aids debugging as, if the OpenMP spawning sections are disabled, the
calculation should be independent of the number of threads used.