Fix OpenMP race condition in redistribution of excips.
The number of particles left on a processor after the redistribution of excips
was updated by all OpenMP threads at the same time rather than using a
reduction.  This is bad, as the number of particles is then not consistent with
the population actually on the processor.  Oops.

Only affects hybrid CCMC calculations.