Use D0_population rather than D0_population_cycle in select_ref_det
D0_population is averaged over the report loop (as claimed in the output
message) whilst D0_population_cycle is not.  Further, D0_population is
available on all processors (as it's reduced in update_energy_estimators)
whereas D0_population_cycle is only stored on the processor containing
the reference.

Apart from a consistency issue, this made some processors think the
reference should be updated and others not, leading to a deadlock...