Correctly detect best determinant is the current reference in parallel
The 'new' best determinant is not available on all processors until it
is broadcast so we can't compare to it.  Instead, send the best
determinant on the processor only if the best determinant is not the
reference, rather than attempt to detect if the best determinat is not
the reference.