Replace C++ MurmurHash2 with C port.
This is just a bit cleaner to link to and removes the DET_SIZE etc
hackery in MurmurHash.cpp which wasn't necessary.

Given we just (effectively) use the hash for small hash tables, a 32-bit
hash is more than sufficient.

The Markov chain for DET_SIZE=32 is unchanged but probably is different
for DET_SIZE=64 (which is a bit broken anyway--see following commit).
Either way, results remain statistically correct as only the load
balancing will be affected.

Note we now require a C rather than C++ compiler.