Teach select_cluster to produce clusters within a range of sizes.
select_cluster used to select a cluster randomly of from a 0-cluster up
to a cluster of a specific number of excitors.  It's useful to handle
small clusters separately, so select_cluster needs to be flexible enough
to generate random clusters containing a number of excitors between two
provided limits.