BUG FIX: correct core energy when freezing orbitals.
In certain situations (detailed below), the same exchange integral was
being included multiple times in the core energy when freezing core
orbitals, leading to a rigid offset of all energies.

This was only the case if the integral occurred multiple times in the
integral file in molecular calculations`, specifically if
< i j | i j > and < i i | j j > both occurred or < i i | j j > occurred
multiple times (e.g. < j j | i i > was also present).