Increase memory usage in ifciqmc/Ne-RHF-aug-cc-pVDZ/hande_read.in test.
It turns out that the spawn_t object was already pretty full.  Using
POP_SIZE=64 doubles the size of some objects and (as we specified
a memory limit in the test) hence halves the number of elements in those
objects.  This caused the test to run out of memory with POP_SIZE=64.

Doubling the memory available (to 2 MB!) for the spawn_t object fixes
this.  We really should add code to do checking of the available space
in spawn_t rather than blindly adding to it...