Option to disable code depending upon the HDF5 library.
Compiling HDF5 with Fortran 2003 support on some architectures (e.g.
Cray!) can be painful.  Hence for initial porting and testing, it is
helpful if we can optionally disable this so any other problems can be
resolved in parallel.  Still, I don't recommend using this option for
production work, as there the restart functionality is very, very
useful.