Bug fix (Hilbert space MC): unbiased determinant generation.
In order to estimate the size of a (truncated) CI space, we generate random excitations of
a reference determinant.  Previously, we discarded attempts that excited from or to the
same orbital multiple times.  This caused a bias towards low excitation levels (as the
chance of selecting the same orbital again increased withe the number of attempts).

A far better approach is to only select from orbitals which have not been previously
selected, which is both faster and ensures that determinants are selected uniformly.