Gracefully handle data server not being able to read file.
If user A is running the extract_test_data server and user B tries to
connect to it and extract data from a file, then the data server returns
an empty string if an exception is encountered in parsing the file.
This arises if user A can't read user B's file.  The simplest (and most
robust) way to handle this is to simply attempt the extraction ourselves
if the data server doesn't return anything.