Correct selection probabilities for clusters in the CCMC full non-composite algorithm.
The selection probability includes the number of clusters that are generated by
a processor in a single timestep.  This was correctly set for the original
algorithm but not for the full non-composite algorithm where the number of
attempts is divided between stochastic selection of composite clusters and
deterministic selection of single excitors.